requirements.txt
Flask-MySQLdb==1.0.1
Flask-WTF==1.1.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
WTForms==3.0.1
Flask==3.0.3
mysql-connector-python==9.0.0
Werkzeug==3.0.3
